Continue moving transfer to updated location and remove unused items
